write keyvault and cosmosdb integration test
