+++ azure-spring-boot-tests/azure-spring-boot-test-core/pom.xml
@@ -5,12 +5,12 @@
     <parent>
         <artifactId>azure-spring-boot-tests</artifactId>
         <groupId>com.microsoft.azure</groupId>
-        <version>2.1.8-SNAPSHOT</version>
+        <version>2.2.0-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>azure-spring-boot-tests-core</artifactId>
+    <artifactId>azure-spring-boot-test-core</artifactId>
 
     <properties>
         <project.rootdir>${project.basedir}/../..</project.rootdir>
@@ -19,15 +19,20 @@
     <dependencies>
         <dependency>
             <groupId>com.microsoft.azure</groupId>
-            <artifactId>azure-mgmt-keyvault</artifactId>
+            <artifactId>azure</artifactId>
         </dependency>
         <dependency>
             <groupId>com.microsoft.azure</groupId>
-            <artifactId>azure-client-authentication</artifactId>
+            <artifactId>azure-documentdb</artifactId>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <scope>provided</scope>
         </dependency>
     </dependencies>
