+++ azure-spring-boot-tests/azure-spring-boot-test-core/src/main/java/com/microsoft/azure/mgmt/Constants.java
@@ -1,0 +1,5 @@
+package com.microsoft.azure.mgmt;
+
+public class Constants {
+    public final static String TEST_RESOURCE_GROUP_NAME = "spring-boot-test-rg";
+}
