+++ azure-spring-boot-tests/azure-spring-boot-test-cosmosdb/pom.xml
@@ -10,7 +10,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>azure-spring-boot-test-keyvault</artifactId>
+    <artifactId>azure-spring-boot-test-cosmosdb</artifactId>
 
     <properties>
         <project.rootdir>${project.basedir}/../..</project.rootdir>
@@ -19,16 +19,25 @@
     <dependencies>
         <dependency>
             <groupId>com.microsoft.azure</groupId>
-            <artifactId>azure-spring-boot-test-core</artifactId>
+            <artifactId>azure-cosmosdb-spring-boot-starter</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>io.projectreactor.netty</groupId>
+            <artifactId>reactor-netty</artifactId>
+            <version>0.8.3.RELEASE</version>
+        </dependency>
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <scope>provided</scope>
         </dependency>
         <dependency>
             <groupId>com.microsoft.azure</groupId>
-            <artifactId>azure-keyvault-secrets-spring-boot-starter</artifactId>
+            <artifactId>azure-spring-boot-test-core</artifactId>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-test</artifactId>
-            <scope>test</scope>
         </dependency>
     </dependencies>
 
@@ -44,5 +53,4 @@
             </plugin>
         </plugins>
     </build>
-
 </project>
