+++ azure-spring-boot-tests/azure-spring-boot-test-cosmosdb/src/test/java/com/microsoft/azure/test/cosmosdb/DummyApp.java
@@ -1,0 +1,9 @@
+package com.microsoft.azure.test.cosmosdb;
+
+import org.springframework.boot.SpringBootConfiguration;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+
+@SpringBootConfiguration
+@EnableAutoConfiguration
+public class DummyApp {
+}
