+++ azure-spring-boot-tests/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <groupId>com.microsoft.azure</groupId>
         <artifactId>azure-spring-boot-parent</artifactId>
-        <version>2.1.8-SNAPSHOT</version>
+        <version>2.2.0-SNAPSHOT</version>
         <relativePath>../azure-spring-boot-parent/pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
@@ -17,22 +17,39 @@
     <description>Tests for Azure Spring Boot</description>
 
     <modules>
-        <module>azure-spring-boot-tests-core</module>
+        <module>azure-spring-boot-test-core</module>
         <module>azure-spring-boot-test-keyvault</module>
+        <module>azure-spring-boot-test-cosmosdb</module>
     </modules>
 
     <properties>
-        <spring.boot.version>2.1.7.RELEASE</spring.boot.version>
-        <azure.spring.boot.version>2.1.8-SNAPSHOT</azure.spring.boot.version>
+        <spring.boot.version>2.2.0.M6</spring.boot.version>
+        <azure.spring.boot.version>2.2.0-SNAPSHOT</azure.spring.boot.version>
         <azure.mgmt.keyvault.version>1.26.0</azure.mgmt.keyvault.version>
+        <azure.documentdb.version>1.9.4</azure.documentdb.version>
     </properties>
 
     <dependencyManagement>
         <dependencies>
             <dependency>
+                <groupId>com.microsoft.azure</groupId>
+                <artifactId>azure-spring-boot-test-core</artifactId>
+                <version>${azure.spring.boot.version}</version>
+            </dependency>
+            <dependency>
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot</artifactId>
                 <version>${spring.boot.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.microsoft.azure</groupId>
+                <artifactId>azure</artifactId>
+                <version>${azure.mgmt.keyvault.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.microsoft.azure</groupId>
+                <artifactId>azure-documentdb</artifactId>
+                <version>${azure.documentdb.version}</version>
             </dependency>
             <dependency>
                 <groupId>com.microsoft.azure</groupId>
@@ -41,17 +58,24 @@
             </dependency>
             <dependency>
                 <groupId>com.microsoft.azure</groupId>
-                <artifactId>azure-mgmt-keyvault</artifactId>
-                <version>${azure.mgmt.keyvault.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>com.microsoft.azure</groupId>
-                <artifactId>azure-spring-boot-tests-core</artifactId>
+                <artifactId>azure-cosmosdb-spring-boot-starter</artifactId>
                 <version>${azure.spring.boot.version}</version>
             </dependency>
         </dependencies>
 
     </dependencyManagement>
 
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-surefire-plugin</artifactId>
+                <version>2.22.0</version>
+                <configuration>
+                    <systemPropertiesFile>config.properties</systemPropertiesFile>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
 
 </project>
